Security

+++ web_dynamic/api/courses.py
@@ -15,18 +15,21 @@ from web_dynamic.api import api_views
 @api_views.route('/courses', strict_slashes=False, methods=['GET'])
 def get_courses():
     """view function to get users courses"""
-    user_id = session.get('user_id')
+    if not g.username:
+        return make_response(jsonify({'error': 'invalid details user not in session'}), 400)
+
+    user_id = g.user_id
     try:
-        semester = int(session.get('semester'))
-        year = int(session.get('year'))
+        semester = int(g.semester)
+        year = int(g.year)
     except(Exception):
         return make_response(jsonify({'error': 'invalid details'}), 400)
 
     if user_id and semester and year:
-        courses = models.storage.all(Course)
-        c_list = [course.to_dict() for course in courses.values()
-                  if course.user_id == user_id
-                  and course.semester == semester and course.year == year]
+        courses = models.storage.get(User, user_id).courses
+        c_list = [course.to_dict() for course in courses
+                  if course.semester == semester 
+                  and course.year == year]
 
         return make_response(jsonify(c_list), 200)
     return make_response(jsonify({'error': 'invalid details'}), 400)
@@ -35,9 +38,12 @@ def get_courses():
 @api_views.route('/courses', strict_slashes=False, methods=['POST'])
 def create_courses():
     """view function to create users courses"""
-    user_id = session.get('user_id')
-    semester = session.get('semester')
-    year = session.get('year')
+    if not g.username:
+        return make_response(jsonify({'error': 'invalid details user not in session'}), 400)
+
+    user_id = g.user_id
+    semester = g.semester
+    year = g.year
 
     name = request.form['name']
     description = request.form['description']
@@ -56,7 +62,10 @@ def create_courses():
 @api_views.route('/set', strict_slashes=False, methods=['PUT'])
 def set_default_sch_info():
     """set the default semester and year to display"""
-    user_id = session.get('user_id')
+    if not g.username:
+        return make_response(jsonify({'error': 'invalid details'}), 400)
+
+    user_id = g.user_id
     user = models.storage.get(User, user_id)
 
     if not user:
@@ -69,23 +78,26 @@ def set_default_sch_info():
     session['semester'] = user.current_semester = data.get('semester')
 
     models.storage.save()
-    u_d = user.to_dict()
-    current_details = {'year': u_d['current_year'],
-                       'semester': u_d['current_semester']}
+    user_dict = user.to_dict()
+    current_details = {'year': user_dict['current_year'],
+                       'semester': user_dict['current_semester']}
     return make_response(jsonify(current_details), 200)
 
 
 @api_views.route('/set', strict_slashes=False, methods=['GET'])
 def get_default_sch_info():
     """get the current selected school info"""
-    user_id = session.get('user_id')
-    semester = session.get('semester')
-    year = session.get('year')
-    user = models.storage.get(User, user_id)
+    if not g.username:
+        return make_response(jsonify({'error': 'invalid details'}), 400)
 
-    if user_id and semester and year:
+    user_id = g.user_id
+    username = g.username
+    semester = g.semester
+    year = g.year
+    
+    if user_id and semester and year and username:
         info_dict = {'user_id': user_id, 'semester': semester,
-                     'year': year, 'username': user.username}
+                     'year': year, 'username': username}
         return make_response(jsonify(info_dict), 200)
     return make_response(jsonify({}), 200)
 
@@ -94,10 +106,25 @@ def get_default_sch_info():
                  strict_slashes=False, methods=['DELETE'])
 def delete_course(course_id=None):
     """delete course based on id"""
+    if not g.username:
+        return make_response(jsonify({'error': 'invalid details'}), 400)
+
+    user_id = g.user_id
     course = models.storage.get(Course, course_id)
 
-    if course:
+    if course and course.user_id == user_id:
         models.storage.delete(course)
         models.storage.save()
         return make_response(jsonify({}), 200)
     return make_response(jsonify({}), 400)
+
+
+@api_views.before_request
+def load_user():
+    g.username = None
+
+    if 'user' in session:
+        g.username = session.get('user')
+        g.user_id = session.get('user_id')
+        g.year = session.get('year')
+        g.semester = session.get('semester')
